requirements.txt
#AI/LLM
langchain_openai
langchain_anthropic
langchain_google_genai
langchain-tavily
pydantic
langgraph
langchain
fastapi
openai-whisper
uvicorn
websockets
torch
pyttsx3
gtts
python-dotenv
ipython
pydantic[email]
langchain-google-community[gmail]

# Database dependencies
sqlalchemy
psycopg2
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart